@@ -1,6 +1,6 @@
 {
   "name": "lemmy-ui",
-  "version": "0.19.0-beta.8",
+  "version": "0.19.0-rc.13",
   "description": "An isomorphic UI for lemmy",
   "repository": "https://github.com/LemmyNet/lemmy-ui",
   "license": "AGPL-3.0",
@@ -35,24 +35,24 @@
     ]
   },
   "dependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-proposal-decorators": "^7.21.5",
     "@babel/plugin-transform-runtime": "^7.21.5",
     "@babel/plugin-transform-typescript": "^7.21.5",
-    "@babel/preset-env": "7.21.5",
+    "@babel/preset-env": "^7.23.3",
     "@babel/preset-typescript": "^7.21.5",
     "@babel/runtime": "^7.21.5",
     "@emoji-mart/data": "^1.1.0",
     "@shortcm/qr-image": "^9.0.2",
     "autosize": "^6.0.1",
-    "babel-loader": "^9.1.2",
+    "babel-loader": "^9.1.3",
     "babel-plugin-inferno": "^6.6.0",
     "bootstrap": "^5.3.1",
     "check-password-strength": "^2.0.7",
     "classnames": "^2.3.1",
     "clean-webpack-plugin": "^4.0.0",
-    "cookie": "^0.5.0",
+    "cookie": "^0.6.0",
     "copy-webpack-plugin": "^11.0.0",
-    "cross-fetch": "^4.0.0",
     "css-loader": "^6.7.3",
     "date-fns": "^2.30.0",
     "emoji-mart": "^5.4.0",
@@ -69,14 +69,15 @@
     "inferno-i18next-dess": "0.0.2",
     "inferno-router": "^8.2.2",
     "inferno-server": "^8.2.2",
-    "jwt-decode": "^3.1.2",
-    "lemmy-js-client": "^0.19.0-rc.13",
+    "jwt-decode": "^4.0.0",
+    "lemmy-js-client": "0.19.0-rc.19",
     "lodash.isequal": "^4.5.0",
     "markdown-it": "^13.0.1",
     "markdown-it-bidi": "^0.1.0",
     "markdown-it-container": "^3.0.0",
     "markdown-it-emoji": "^2.0.2",
     "markdown-it-footnote": "^3.0.3",
+    "markdown-it-highlightjs": "^4.0.1",
     "markdown-it-html5-embed": "^1.0.0",
     "markdown-it-ruby": "^0.1.1",
     "markdown-it-sub": "^1.0.0",
@@ -90,26 +91,25 @@
     "sass-loader": "^13.3.2",
     "serialize-javascript": "^6.0.1",
     "service-worker-webpack": "^1.0.0",
-    "sharp": "^0.32.4",
+    "sharp": "^0.32.6",
     "tippy.js": "^6.3.7",
     "toastify-js": "^1.12.0",
     "tributejs": "^5.1.3",
-    "webpack": "5.88.2",
+    "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4",
     "webpack-node-externals": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.5",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/core": "^7.23.3",
     "@types/autosize": "^4.0.0",
     "@types/bootstrap": "^5.2.6",
-    "@types/cookie": "^0.5.1",
+    "@types/cookie": "^0.6.0",
     "@types/express": "^4.17.17",
     "@types/html-to-text": "^9.0.0",
     "@types/lodash.isequal": "^4.5.6",
-    "@types/markdown-it": "^12.2.3",
+    "@types/markdown-it": "^13.0.7",
     "@types/markdown-it-container": "^2.0.6",
-    "@types/node": "^20.4.5",
+    "@types/node": "^20.10.0",
     "@types/path-browserify": "^1.0.0",
     "@types/sanitize-html": "^2.9.0",
     "@types/serialize-javascript": "^5.0.1",
@@ -121,7 +121,7 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.0",
     "import-sort-style-module": "^6.0.0",
-    "lint-staged": "^13.2.3",
+    "lint-staged": "^15.1.0",
     "prettier": "^3.0.0",
     "prettier-plugin-import-sort": "^0.0.7",
     "prettier-plugin-organize-imports": "^3.2.3",
@@ -131,7 +131,7 @@
     "style-loader": "^3.3.2",
     "terser": "^5.19.2",
     "typescript": "^5.1.6",
-    "typescript-language-server": "^3.3.2",
+    "typescript-language-server": "^4.1.2",
     "webpack-bundle-analyzer": "^4.9.0",
     "webpack-dev-server": "4.15.1"
   },
