@@ -2,7 +2,6 @@ import { initializeSite, setupDateFns } from "@utils/app";
 import { hydrate } from "inferno-hydrate";
 import { BrowserRouter } from "inferno-router";
 import { App } from "../shared/components/app/app";
-import { UserService } from "../shared/services";
 
 import "bootstrap/js/dist/collapse";
 import "bootstrap/js/dist/dropdown";
@@ -15,7 +14,7 @@ async function startClient() {
 
   const wrapper = (
     <BrowserRouter>
-      <App user={UserService.Instance.myUserInfo} />
+      <App />
     </BrowserRouter>
   );
 
