@@ -1,8 +1,6 @@
 import { initializeSite, isAuthPath } from "@utils/app";
 import { getHttpBaseInternal } from "@utils/env";
 import { ErrorPageData } from "@utils/types";
-import * as cookie from "cookie";
-import fetch from "cross-fetch";
 import type { Request, Response } from "express";
 import { StaticRouter, matchPath } from "inferno-router";
 import { renderToString } from "inferno-server";
@@ -21,24 +19,19 @@ import {
 import { createSsrHtml } from "../utils/create-ssr-html";
 import { getErrorPageData } from "../utils/get-error-page-data";
 import { setForwardedHeaders } from "../utils/set-forwarded-headers";
+import { getJwtCookie } from "../utils/has-jwt-cookie";
 
 export default async (req: Request, res: Response) => {
   try {
     const activeRoute = routes.find(route => matchPath(req.path, route));
 
     const headers = setForwardedHeaders(req.headers);
+    const auth = getJwtCookie(req.headers);
 
     const client = wrapClient(
-      new LemmyHttp(getHttpBaseInternal(), { fetchFunction: fetch, headers }),
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
     );
 
-    const auth = req.headers.cookie
-      ? cookie.parse(req.headers.cookie).jwt
-      : undefined;
-
-    if (auth) {
-      client.setHeaders({ Authorization: `Bearer ${auth}` });
-    }
     const { path, url, query } = req;
 
     // Get site data first
@@ -49,7 +42,10 @@ export default async (req: Request, res: Response) => {
     let errorPageData: ErrorPageData | undefined = undefined;
     let try_site = await client.getSite();
 
-    if (try_site.state === "failed" && try_site.msg === "not_logged_in") {
+    if (
+      try_site.state === "failed" &&
+      try_site.err.message === "not_logged_in"
+    ) {
       console.error(
         "Incorrect JWT token, skipping auth so frontend can remove jwt cookie",
       );
@@ -71,10 +67,10 @@ export default async (req: Request, res: Response) => {
 
       if (site && activeRoute?.fetchInitialData) {
         const initialFetchReq: InitialFetchRequest = {
-          client,
           path,
           query,
           site,
+          headers,
         };
 
         routeData = await activeRoute.fetchInitialData(initialFetchReq);
@@ -85,22 +81,23 @@ export default async (req: Request, res: Response) => {
       }
     } else if (try_site.state === "failed") {
       res.status(500);
-      errorPageData = getErrorPageData(new Error(try_site.msg), site);
+      errorPageData = getErrorPageData(new Error(try_site.err.message), site);
     }
 
     const error = Object.values(routeData).find(
-      res => res.state === "failed" && res.msg !== "couldnt_find_object", // TODO: find a better way of handling errors
+      res =>
+        res.state === "failed" && res.err.message !== "couldnt_find_object", // TODO: find a better way of handling errors
     ) as FailedRequestState | undefined;
 
     // Redirect to the 404 if there's an API error
     if (error) {
-      console.error(error.msg);
+      console.error(error.err);
 
-      if (error.msg === "instance_is_private") {
+      if (error.err.message === "instance_is_private") {
         return res.redirect(`/signup`);
       } else {
         res.status(500);
-        errorPageData = getErrorPageData(new Error(error.msg), site);
+        errorPageData = getErrorPageData(new Error(error.err.message), site);
       }
     }
 
@@ -113,7 +110,7 @@ export default async (req: Request, res: Response) => {
 
     const wrapper = (
       <StaticRouter location={url} context={isoData}>
-        <App user={site?.my_user} />
+        <App />
       </StaticRouter>
     );
 
