@@ -1,24 +1,19 @@
-import { getHttpBaseExternal, getHttpBaseInternal } from "@utils/env";
-import fetch from "cross-fetch";
+import { getHttpBaseInternal } from "@utils/env";
 import type { Request, Response } from "express";
 import { LemmyHttp } from "lemmy-js-client";
 import { wrapClient } from "../../shared/services/HttpService";
 import generateManifestJson from "../utils/generate-manifest-json";
-import { setForwardedHeaders } from "../utils/set-forwarded-headers";
 
 let manifest: Awaited<ReturnType<typeof generateManifestJson>> | undefined =
   undefined;
 
-export default async (req: Request, res: Response) => {
-  if (!manifest || manifest.start_url !== getHttpBaseExternal()) {
-    const headers = setForwardedHeaders(req.headers);
-    const client = wrapClient(
-      new LemmyHttp(getHttpBaseInternal(), { fetchFunction: fetch, headers }),
-    );
+export default async (_req: Request, res: Response) => {
+  if (!manifest) {
+    const client = wrapClient(new LemmyHttp(getHttpBaseInternal()));
     const site = await client.getSite();
 
     if (site.state === "success") {
-      manifest = await generateManifestJson(site.data);
+      manifest = await generateManifestJson(site.data.site_view.site);
     } else {
       res.sendStatus(500);
       return;
