@@ -1,6 +1,6 @@
 import * as crypto from "crypto";
 import type { NextFunction, Request, Response } from "express";
-import { hasJwtCookie } from "./utils/has-jwt-cookie";
+import { getJwtCookie } from "./utils/has-jwt-cookie";
 
 export function setDefaultCsp({
   res,
@@ -16,7 +16,7 @@ export function setDefaultCsp({
     `default-src 'self';
      manifest-src *;
      connect-src *;
-     img-src * data:;
+     img-src * data: blob:;
      script-src 'self' 'nonce-${res.locals.cspNonce}';
      style-src 'self' 'unsafe-inline';
      form-action 'self';
@@ -52,7 +52,7 @@ export function setCacheControl(
     // Static content gets cached publicly for a day
     caching = "public, max-age=86400";
   } else {
-    if (hasJwtCookie(req)) {
+    if (getJwtCookie(req.headers)) {
       caching = "private";
     } else {
       caching = "public, max-age=60";
