@@ -1,7 +1,6 @@
 import { readFile } from "fs/promises";
-import { GetSiteResponse } from "lemmy-js-client";
+import { Site } from "lemmy-js-client";
 import path from "path";
-import sharp from "sharp";
 import { fetchIconPng } from "./fetch-icon-png";
 
 const iconSizes = [72, 96, 128, 144, 152, 192, 384, 512];
@@ -13,13 +12,7 @@ const defaultLogoPathDirectory = path.join(
   "icons",
 );
 
-export default async function ({
-  my_user,
-  site_view: {
-    site,
-    local_site: { community_creation_admin_only },
-  },
-}: GetSiteResponse) {
+export default async function (site: Site) {
   const icon = site.icon ? await fetchIconPng(site.icon) : null;
 
   return {
@@ -38,6 +31,7 @@ export default async function ({
         ).then(buf => buf.toString("base64"));
 
         if (icon) {
+          const sharp = (await import("sharp")).default;
           src = await sharp(icon)
             .resize(size, size)
             .png()
@@ -72,19 +66,7 @@ export default async function ({
         short_name: "Create Post",
         description: "Create a post.",
       },
-    ].concat(
-      my_user?.local_user_view.local_user.admin ||
-        !community_creation_admin_only
-        ? [
-            {
-              name: "Create Community",
-              url: "/create_community",
-              short_name: "Create Community",
-              description: "Create a community",
-            },
-          ]
-        : [],
-    ),
+    ],
     related_applications: [
       {
         platform: "f-droid",
