@@ -3,10 +3,9 @@ import { dataBsTheme } from "@utils/browser";
 import { Component, RefObject, createRef, linkEvent } from "inferno";
 import { Provider } from "inferno-i18next-dess";
 import { Route, Switch } from "inferno-router";
-import { MyUserInfo } from "lemmy-js-client";
 import { IsoDataOptionalSite } from "../../interfaces";
 import { routes } from "../../routes";
-import { FirstLoadService, I18NextService, UserService } from "../../services";
+import { FirstLoadService, I18NextService } from "../../services";
 import AuthGuard from "../common/auth-guard";
 import ErrorGuard from "../common/error-guard";
 import { ErrorPage } from "./error-page";
@@ -15,15 +14,12 @@ import { Navbar } from "./navbar";
 import "./styles.scss";
 import { Theme } from "./theme";
 import AnonymousGuard from "../common/anonymous-guard";
+import { CodeTheme } from "./code-theme";
 
-interface AppProps {
-  user?: MyUserInfo;
-}
-
-export class App extends Component<AppProps, any> {
+export class App extends Component<any, any> {
   private isoData: IsoDataOptionalSite = setIsoData(this.context);
   private readonly mainContentRef: RefObject<HTMLElement>;
-  constructor(props: AppProps, context: any) {
+  constructor(props: any, context: any) {
     super(props, context);
     this.mainContentRef = createRef();
   }
@@ -32,8 +28,6 @@ export class App extends Component<AppProps, any> {
     event.preventDefault();
     this.mainContentRef.current?.focus();
   }
-
-  user = UserService.Instance.myUserInfo;
 
   render() {
     const siteRes = this.isoData.site_res;
@@ -45,7 +39,7 @@ export class App extends Component<AppProps, any> {
           <div
             id="app"
             className="lemmy-site"
-            data-bs-theme={dataBsTheme(this.props.user)}
+            data-bs-theme={dataBsTheme(siteRes)}
           >
             <button
               type="button"
@@ -55,7 +49,10 @@ export class App extends Component<AppProps, any> {
               {I18NextService.i18n.t("jump_to_content", "Jump to content")}
             </button>
             {siteView && (
-              <Theme defaultTheme={siteView.local_site.default_theme} />
+              <>
+                <Theme defaultTheme={siteView.local_site.default_theme} />
+                <CodeTheme />
+              </>
             )}
             <Navbar siteRes={siteRes} />
             <div className="mt-4 p-0 fl-1">
