@@ -14,6 +14,7 @@ import {
   EditSite,
   GetFederatedInstancesResponse,
   GetSiteResponse,
+  LemmyHttp,
   PersonView,
 } from "lemmy-js-client";
 import { InitialFetchRequest } from "../../interfaces";
@@ -24,6 +25,7 @@ import {
   HttpService,
   LOADING_REQUEST,
   RequestState,
+  wrapClient,
 } from "../../services/HttpService";
 import { toast } from "../../toast";
 import { HtmlTags } from "../common/html-tags";
@@ -35,6 +37,7 @@ import { ExternalAuthForm } from "./external-auth-form";
 import RateLimitForm from "./rate-limit-form";
 import { SiteForm } from "./site-form";
 import { TaglineForm } from "./tagline-form";
+import { getHttpBaseInternal } from "../../utils/env";
 
 type AdminSettingsData = RouteDataResponse<{
   bannedRes: BannedPersonsResponse;
@@ -92,8 +95,11 @@ export class AdminSettings extends Component<any, AdminSettingsState> {
   }
 
   static async fetchInitialData({
-    client,
+    headers,
   }: InitialFetchRequest): Promise<AdminSettingsData> {
+    const client = wrapClient(
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
+    );
     return {
       bannedRes: await client.getBannedPersons(),
       instancesRes: await client.getFederatedInstances(),
@@ -387,7 +393,7 @@ export class AdminSettings extends Component<any, AdminSettingsState> {
   async handleDeleteEmoji(form: DeleteCustomEmoji) {
     const res = await HttpService.client.deleteCustomEmoji(form);
     if (res.state === "success") {
-      removeFromEmojiDataModel(res.data.id);
+      removeFromEmojiDataModel(form.id);
     }
   }
 
