@@ -1,4 +1,4 @@
-import { editRegistrationApplication, myAuth, setIsoData } from "@utils/app";
+import { editRegistrationApplication, setIsoData } from "@utils/app";
 import { randomStr } from "@utils/helpers";
 import { RouteDataResponse } from "@utils/types";
 import classNames from "classnames";
@@ -6,6 +6,7 @@ import { Component, linkEvent } from "inferno";
 import {
   ApproveRegistrationApplication,
   GetSiteResponse,
+  LemmyHttp,
   ListRegistrationApplicationsResponse,
   RegistrationApplicationView,
 } from "lemmy-js-client";
@@ -17,12 +18,15 @@ import {
   HttpService,
   LOADING_REQUEST,
   RequestState,
+  wrapClient,
 } from "../../services/HttpService";
 import { setupTippy } from "../../tippy";
 import { HtmlTags } from "../common/html-tags";
 import { Spinner } from "../common/icon";
 import { Paginator } from "../common/paginator";
 import { RegistrationApplication } from "../common/registration-application";
+import { UnreadCounterService } from "../../services";
+import { getHttpBaseInternal } from "../../utils/env";
 
 enum UnreadOrAll {
   Unread,
@@ -206,10 +210,13 @@ export class RegistrationApplications extends Component<
   }
 
   static async fetchInitialData({
-    client,
+    headers,
   }: InitialFetchRequest): Promise<RegistrationApplicationsData> {
+    const client = wrapClient(
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
+    );
     return {
-      listRegistrationApplicationsResponse: myAuth()
+      listRegistrationApplicationsResponse: headers["Authorization"]
         ? await client.listRegistrationApplications({
             unread_only: true,
             page: 1,
@@ -234,15 +241,18 @@ export class RegistrationApplications extends Component<
   }
 
   async handleApproveApplication(form: ApproveRegistrationApplication) {
-    const approveRes = await HttpService.client.approveRegistrationApplication(
-      form,
-    );
+    const approveRes =
+      await HttpService.client.approveRegistrationApplication(form);
     this.setState(s => {
       if (s.appsRes.state === "success" && approveRes.state === "success") {
         s.appsRes.data.registration_applications = editRegistrationApplication(
           approveRes.data.registration_application,
           s.appsRes.data.registration_applications,
         );
+        if (this.state.unreadOrAll === UnreadOrAll.Unread) {
+          this.refetch();
+          UnreadCounterService.Instance.updateApplications();
+        }
       }
       return s;
     });
