@@ -13,6 +13,7 @@ import {
   CommentReportResponse,
   CommentReportView,
   GetSiteResponse,
+  LemmyHttp,
   ListCommentReports,
   ListCommentReportsResponse,
   ListPostReports,
@@ -39,6 +40,7 @@ import {
   EMPTY_REQUEST,
   LOADING_REQUEST,
   RequestState,
+  wrapClient,
 } from "../../services/HttpService";
 import { CommentReport } from "../comment/comment-report";
 import { HtmlTags } from "../common/html-tags";
@@ -46,6 +48,8 @@ import { Spinner } from "../common/icon";
 import { Paginator } from "../common/paginator";
 import { PostReport } from "../post/post-report";
 import { PrivateMessageReport } from "../private_message/private-message-report";
+import { UnreadCounterService } from "../../services";
+import { getHttpBaseInternal } from "../../utils/env";
 
 enum UnreadOrAll {
   Unread,
@@ -534,8 +538,11 @@ export class Reports extends Component<any, ReportsState> {
   }
 
   static async fetchInitialData({
-    client,
+    headers,
   }: InitialFetchRequest): Promise<ReportsData> {
+    const client = wrapClient(
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
+    );
     const unresolved_only = true;
     const page = 1;
     const limit = fetchLimit;
@@ -600,9 +607,8 @@ export class Reports extends Component<any, ReportsState> {
 
     if (amAdmin()) {
       this.setState({
-        messageReportsRes: await HttpService.client.listPrivateMessageReports(
-          form,
-        ),
+        messageReportsRes:
+          await HttpService.client.listPrivateMessageReports(form),
       });
     }
   }
@@ -610,16 +616,28 @@ export class Reports extends Component<any, ReportsState> {
   async handleResolveCommentReport(form: ResolveCommentReport) {
     const res = await HttpService.client.resolveCommentReport(form);
     this.findAndUpdateCommentReport(res);
+    if (this.state.unreadOrAll === UnreadOrAll.Unread) {
+      this.refetch();
+      UnreadCounterService.Instance.updateReports();
+    }
   }
 
   async handleResolvePostReport(form: ResolvePostReport) {
     const res = await HttpService.client.resolvePostReport(form);
     this.findAndUpdatePostReport(res);
+    if (this.state.unreadOrAll === UnreadOrAll.Unread) {
+      this.refetch();
+      UnreadCounterService.Instance.updateReports();
+    }
   }
 
   async handleResolvePrivateMessageReport(form: ResolvePrivateMessageReport) {
     const res = await HttpService.client.resolvePrivateMessageReport(form);
     this.findAndUpdatePrivateMessageReport(res);
+    if (this.state.unreadOrAll === UnreadOrAll.Unread) {
+      this.refetch();
+      UnreadCounterService.Instance.updateReports();
+    }
   }
 
   findAndUpdateCommentReport(res: RequestState<CommentReportResponse>) {
