@@ -1,4 +1,9 @@
-import { enableDownvotes, enableNsfw, setIsoData } from "@utils/app";
+import {
+  communityToChoice,
+  enableDownvotes,
+  enableNsfw,
+  setIsoData,
+} from "@utils/app";
 import { getIdFromString, getQueryParams } from "@utils/helpers";
 import type { QueryParams } from "@utils/types";
 import { Choice, RouteDataResponse } from "@utils/types";
@@ -9,6 +14,7 @@ import {
   GetCommunity,
   GetCommunityResponse,
   GetSiteResponse,
+  LemmyHttp,
   ListCommunitiesResponse,
 } from "lemmy-js-client";
 import { InitialFetchRequest, PostFormParams } from "../../interfaces";
@@ -18,10 +24,12 @@ import {
   HttpService,
   RequestState,
   WrappedLemmyHttp,
+  wrapClient,
 } from "../../services/HttpService";
 import { HtmlTags } from "../common/html-tags";
 import { Spinner } from "../common/icon";
 import { PostForm } from "./post-form";
+import { getHttpBaseInternal } from "../../utils/env";
 
 export interface CreatePostProps {
   communityId?: number;
@@ -82,10 +90,9 @@ export class CreatePost extends Component<
       };
 
       if (communityRes?.state === "success") {
-        const communityChoice: Choice = {
-          label: communityRes.data.community_view.community.title,
-          value: communityRes.data.community_view.community.id.toString(),
-        };
+        const communityChoice = communityToChoice(
+          communityRes.data.community_view,
+        );
 
         this.state = {
           ...this.state,
@@ -104,10 +111,7 @@ export class CreatePost extends Component<
       });
       if (res.state === "success") {
         this.setState({
-          selectedCommunityChoice: {
-            label: res.data.community_view.community.title,
-            value: res.data.community_view.community.id.toString(),
-          },
+          selectedCommunityChoice: communityToChoice(res.data.community_view),
           loading: false,
         });
       }
@@ -236,11 +240,14 @@ export class CreatePost extends Component<
   }
 
   static async fetchInitialData({
-    client,
+    headers,
     query: { communityId },
   }: InitialFetchRequest<
     QueryParams<CreatePostProps>
   >): Promise<CreatePostData> {
+    const client = wrapClient(
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
+    );
     const data: CreatePostData = {
       initialCommunitiesRes: await fetchCommunitiesForOptions(client),
       communityResponse: EMPTY_REQUEST,
