@@ -18,7 +18,7 @@ import {
   restoreScrollPosition,
   saveScrollPosition,
 } from "@utils/browser";
-import { debounce, randomStr } from "@utils/helpers";
+import { debounce, getApubName, randomStr } from "@utils/helpers";
 import { isImage } from "@utils/media";
 import { RouteDataResponse } from "@utils/types";
 import autosize from "autosize";
@@ -59,14 +59,12 @@ import {
   GetPost,
   GetPostResponse,
   GetSiteResponse,
+  LemmyHttp,
   LockPost,
   MarkCommentReplyAsRead,
-  MarkPersonMentionAsRead,
-  MarkPostAsRead,
   PostResponse,
   PurgeComment,
   PurgeCommunity,
-  PurgeItemResponse,
   PurgePerson,
   PurgePost,
   RemoveComment,
@@ -74,6 +72,7 @@ import {
   RemovePost,
   SaveComment,
   SavePost,
+  SuccessResponse,
   TransferCommunity,
 } from "lemmy-js-client";
 import { commentTreeMaxDepth } from "../../config";
@@ -88,6 +87,7 @@ import {
   HttpService,
   LOADING_REQUEST,
   RequestState,
+  wrapClient,
 } from "../../services/HttpService";
 import { setupTippy } from "../../tippy";
 import { toast } from "../../toast";
@@ -97,6 +97,7 @@ import { HtmlTags } from "../common/html-tags";
 import { Icon, Spinner } from "../common/icon";
 import { Sidebar } from "../community/sidebar";
 import { PostListing } from "./post-listing";
+import { getHttpBaseInternal } from "../../utils/env";
 
 const commentsShownInterval = 15;
 
@@ -164,7 +165,6 @@ export class Post extends Component<any, PostState> {
     this.handleTransferCommunity = this.handleTransferCommunity.bind(this);
     this.handleFetchChildren = this.handleFetchChildren.bind(this);
     this.handleCommentReplyRead = this.handleCommentReplyRead.bind(this);
-    this.handlePersonMentionRead = this.handlePersonMentionRead.bind(this);
     this.handleBanFromCommunity = this.handleBanFromCommunity.bind(this);
     this.handleBanPerson = this.handleBanPerson.bind(this);
     this.handlePostEdit = this.handlePostEdit.bind(this);
@@ -176,7 +176,6 @@ export class Post extends Component<any, PostState> {
     this.handleSavePost = this.handleSavePost.bind(this);
     this.handlePurgePost = this.handlePurgePost.bind(this);
     this.handleFeaturePost = this.handleFeaturePost.bind(this);
-    this.handleMarkPostAsRead = this.handleMarkPostAsRead.bind(this);
 
     this.state = { ...this.state, commentSectionRef: createRef() };
 
@@ -235,9 +234,12 @@ export class Post extends Component<any, PostState> {
   }
 
   static async fetchInitialData({
-    client,
+    headers,
     path,
   }: InitialFetchRequest): Promise<PostData> {
+    const client = wrapClient(
+      new LemmyHttp(getHttpBaseInternal(), { headers }),
+    );
     const pathSplit = path.split("/");
 
     const pathType = pathSplit.at(1);
@@ -261,9 +263,14 @@ export class Post extends Component<any, PostState> {
       commentsForm.parent_id = id;
     }
 
+    const [postRes, commentsRes] = await Promise.all([
+      client.getPost(postForm),
+      client.getComments(commentsForm),
+    ]);
+
     return {
-      postRes: await client.getPost(postForm),
-      commentsRes: await client.getComments(commentsForm),
+      postRes,
+      commentsRes,
     };
   }
 
@@ -387,7 +394,7 @@ export class Post extends Component<any, PostState> {
                 onAddAdmin={this.handleAddAdmin}
                 onTransferCommunity={this.handleTransferCommunity}
                 onFeaturePost={this.handleFeaturePost}
-                onMarkPostAsRead={this.handleMarkPostAsRead}
+                onMarkPostAsRead={() => {}}
               />
               <div ref={this.state.commentSectionRef} className="mb-2" />
 
@@ -589,7 +596,7 @@ export class Post extends Component<any, PostState> {
             onPurgeComment={this.handlePurgeComment}
             onPurgePerson={this.handlePurgePerson}
             onCommentReplyRead={this.handleCommentReplyRead}
-            onPersonMentionRead={this.handlePersonMentionRead}
+            onPersonMentionRead={() => {}}
             onBanPersonFromCommunity={this.handleBanFromCommunity}
             onBanPerson={this.handleBanPerson}
             onCreateComment={this.handleCreateComment}
@@ -676,7 +683,7 @@ export class Post extends Component<any, PostState> {
             onPurgeComment={this.handlePurgeComment}
             onPurgePerson={this.handlePurgePerson}
             onCommentReplyRead={this.handleCommentReplyRead}
-            onPersonMentionRead={this.handlePersonMentionRead}
+            onPersonMentionRead={() => {}}
             onBanPersonFromCommunity={this.handleBanFromCommunity}
             onBanPerson={this.handleBanPerson}
             onCreateComment={this.handleCreateComment}
@@ -745,6 +752,11 @@ export class Post extends Component<any, PostState> {
   async handleAddModToCommunity(form: AddModToCommunity) {
     const addModRes = await HttpService.client.addModToCommunity(form);
     this.updateModerators(addModRes);
+    if (addModRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(form.added ? "appointed_mod" : "removed_mod"),
+      );
+    }
   }
 
   async handleFollow(form: FollowCommunity) {
@@ -822,7 +834,7 @@ export class Post extends Component<any, PostState> {
 
   async handleEditComment(form: EditComment) {
     const editCommentRes = await HttpService.client.editComment(form);
-    this.findAndUpdateComment(editCommentRes);
+    this.findAndUpdateCommentEdit(editCommentRes);
 
     return editCommentRes;
   }
@@ -830,21 +842,45 @@ export class Post extends Component<any, PostState> {
   async handleDeleteComment(form: DeleteComment) {
     const deleteCommentRes = await HttpService.client.deleteComment(form);
     this.findAndUpdateComment(deleteCommentRes);
+    if (deleteCommentRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.deleted ? "deleted_comment" : "undeleted_comment",
+        ),
+      );
+    }
   }
 
   async handleDeletePost(form: DeletePost) {
     const deleteRes = await HttpService.client.deletePost(form);
     this.updatePost(deleteRes);
+    if (deleteRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(form.deleted ? "deleted_post" : "undeleted_post"),
+      );
+    }
   }
 
   async handleRemovePost(form: RemovePost) {
     const removeRes = await HttpService.client.removePost(form);
     this.updatePost(removeRes);
+    if (removeRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(form.removed ? "removed_post" : "restored_post"),
+      );
+    }
   }
 
   async handleRemoveComment(form: RemoveComment) {
     const removeCommentRes = await HttpService.client.removeComment(form);
     this.findAndUpdateComment(removeCommentRes);
+    if (removeCommentRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.removed ? "removed_comment" : "restored_comment",
+        ),
+      );
+    }
   }
 
   async handleSaveComment(form: SaveComment) {
@@ -860,6 +896,13 @@ export class Post extends Component<any, PostState> {
   async handleFeaturePost(form: FeaturePost) {
     const featureRes = await HttpService.client.featurePost(form);
     this.updatePost(featureRes);
+    if (featureRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.featured ? "featured_post" : "unfeatured_post",
+        ),
+      );
+    }
   }
 
   async handleCommentVote(form: CreateCommentLike) {
@@ -870,11 +913,13 @@ export class Post extends Component<any, PostState> {
   async handlePostVote(form: CreatePostLike) {
     const voteRes = await HttpService.client.likePost(form);
     this.updatePost(voteRes);
+    return voteRes;
   }
 
   async handlePostEdit(form: EditPost) {
     const res = await HttpService.client.editPost(form);
     this.updatePost(res);
+    return res;
   }
 
   async handleCommentReport(form: CreateCommentReport) {
@@ -894,11 +939,25 @@ export class Post extends Component<any, PostState> {
   async handleLockPost(form: LockPost) {
     const lockRes = await HttpService.client.lockPost(form);
     this.updatePost(lockRes);
+    if (lockRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(form.locked ? "locked_post" : "unlocked_post"),
+      );
+    }
   }
 
   async handleDistinguishComment(form: DistinguishComment) {
     const distinguishRes = await HttpService.client.distinguishComment(form);
     this.findAndUpdateComment(distinguishRes);
+    if (distinguishRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.distinguished
+            ? "distinguished_comment"
+            : "undistinguished_comment",
+        ),
+      );
+    }
   }
 
   async handleAddAdmin(form: AddAdmin) {
@@ -906,14 +965,19 @@ export class Post extends Component<any, PostState> {
 
     if (addAdminRes.state === "success") {
       this.setState(s => ((s.siteRes.admins = addAdminRes.data.admins), s));
+      toast(
+        I18NextService.i18n.t(form.added ? "appointed_admin" : "removed_admin"),
+      );
     }
   }
 
   async handleTransferCommunity(form: TransferCommunity) {
-    const transferCommunityRes = await HttpService.client.transferCommunity(
-      form,
-    );
+    const transferCommunityRes =
+      await HttpService.client.transferCommunity(form);
     this.updateCommunityFull(transferCommunityRes);
+    if (transferCommunityRes.state === "success") {
+      toast(I18NextService.i18n.t("transferred_community"));
+    }
   }
 
   async handleFetchChildren(form: GetComments) {
@@ -936,24 +1000,35 @@ export class Post extends Component<any, PostState> {
     this.findAndUpdateCommentReply(readRes);
   }
 
-  async handlePersonMentionRead(form: MarkPersonMentionAsRead) {
-    // TODO not sure what to do here. Maybe it is actually optional, because post doesn't need it.
-    await HttpService.client.markPersonMentionAsRead(form);
-  }
-
-  async handleMarkPostAsRead(form: MarkPostAsRead) {
-    const res = await HttpService.client.markPostAsRead(form);
-    this.updatePost(res);
-  }
-
   async handleBanFromCommunity(form: BanFromCommunity) {
     const banRes = await HttpService.client.banFromCommunity(form);
-    this.updateBan(banRes);
+    this.updateBanFromCommunity(banRes);
+    if (banRes.state === "success" && this.state.postRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.ban ? "banned_from_community" : "unbanned_from_community",
+          {
+            user: getApubName(this.state.postRes.data.post_view.creator),
+            community: getApubName(this.state.postRes.data.post_view.community),
+          },
+        ),
+      );
+    }
   }
 
   async handleBanPerson(form: BanPerson) {
     const banRes = await HttpService.client.banPerson(form);
     this.updateBan(banRes);
+    if (banRes.state === "success" && this.state.postRes.state === "success") {
+      toast(
+        I18NextService.i18n.t(
+          form.ban ? "banned_from_site" : "unbanned_from_site",
+          {
+            user: getApubName(this.state.postRes.data.post_view.creator),
+          },
+        ),
+      );
+    }
   }
 
   updateBanFromCommunity(banRes: RequestState<BanFromCommunityResponse>) {
@@ -1029,7 +1104,7 @@ export class Post extends Component<any, PostState> {
     });
   }
 
-  purgeItem(purgeRes: RequestState<PurgeItemResponse>) {
+  purgeItem(purgeRes: RequestState<SuccessResponse>) {
     if (purgeRes.state === "success") {
       toast(I18NextService.i18n.t("purge_success"));
       this.context.router.history.push(`/`);
@@ -1059,7 +1134,7 @@ export class Post extends Component<any, PostState> {
     });
   }
 
-  findAndUpdateComment(res: RequestState<CommentResponse>) {
+  findAndUpdateCommentEdit(res: RequestState<CommentResponse>) {
     this.setState(s => {
       if (s.commentsRes.state === "success" && res.state === "success") {
         s.commentsRes.data.comments = editComment(
@@ -1067,6 +1142,19 @@ export class Post extends Component<any, PostState> {
           s.commentsRes.data.comments,
         );
         s.finished.set(res.data.comment_view.comment.id, true);
+      }
+      return s;
+    });
+  }
+
+  // No need to set finished on a comment vote, save, etc
+  findAndUpdateComment(res: RequestState<CommentResponse>) {
+    this.setState(s => {
+      if (s.commentsRes.state === "success" && res.state === "success") {
+        s.commentsRes.data.comments = editComment(
+          res.data.comment_view,
+          s.commentsRes.data.comments,
+        );
       }
       return s;
     });
