@@ -14,6 +14,7 @@ import markdown_it_html5_embed from "markdown-it-html5-embed";
 import markdown_it_ruby from "markdown-it-ruby";
 import markdown_it_sub from "markdown-it-sub";
 import markdown_it_sup from "markdown-it-sup";
+import markdown_it_highlightjs from "markdown-it-highlightjs";
 import Renderer from "markdown-it/lib/renderer";
 import Token from "markdown-it/lib/token";
 import { instanceLinkRegex, relTags } from "./config";
@@ -62,10 +63,9 @@ const spoilerConfig = {
 
   render: (tokens: any, idx: any) => {
     const m = tokens[idx].info.trim().match(/^spoiler\s+(.*)$/);
-    const summary = mdToHtmlInline(md.utils.escapeHtml(m[1])).__html;
-
     if (tokens[idx].nesting === 1) {
       // opening tag
+      const summary = mdToHtmlInline(md.utils.escapeHtml(m[1])).__html;
       return `<details><summary> ${summary} </summary>\n`;
     } else {
       // closing tag
@@ -170,6 +170,7 @@ export function setupMarkdown() {
     .use(markdown_it_footnote)
     .use(markdown_it_html5_embed, html5EmbedConfig)
     .use(markdown_it_container, "spoiler", spoilerConfig)
+    .use(markdown_it_highlightjs, { inline: true })
     .use(markdown_it_ruby)
     .use(localInstanceLinkParser)
     .use(markdown_it_bidi);
@@ -183,6 +184,7 @@ export function setupMarkdown() {
     .use(markdown_it_footnote)
     .use(markdown_it_html5_embed, html5EmbedConfig)
     .use(markdown_it_container, "spoiler", spoilerConfig)
+    .use(markdown_it_highlightjs, { inline: true })
     .use(localInstanceLinkParser)
     .use(markdown_it_bidi)
     // .use(markdown_it_emoji, {
