@@ -7,8 +7,8 @@ export default function getExternalHost() {
         ["1234", "1235"].includes(window.location.port)
           ? ":8536"
           : window.location.port === ""
-          ? ""
-          : `:${window.location.port}`
+            ? ""
+            : `:${window.location.port}`
       }`
     : process.env.LEMMY_UI_LEMMY_EXTERNAL_HOST || testHost;
 }
